Fix checking effect_handle issue and build script.
Before, it didn't check whethere it is null or not, now it checks.
Before checking architecture on spec file has some problem, so it build as emulator.
So now it checkes whether it has emulator option on or not.

+++ src/haptic/haptic.c
@@ -179,7 +179,7 @@ static void haptic_name_owner_changed(const char *sender, void *data)
 		return;
 
 	for (n = info->handle_list; n; n = n->next) {
-		handle = n->data;
+		handle = (int)n->data;
 		h_ops->stop_device(handle);
 		h_ops->close_device(handle);
 	}
